products orders added

+++ src/actions/auth.actions.js
@@ -58,9 +58,22 @@ export  const isUserLoggedIn = () =>{
 
 export const signOut=()=>{
     return async dispatch =>{
-        localStorage.clear()
         dispatch({
             type:authConstant.LOGOUT_REQUEST
         })
+        const res=await axios.post(`/admin/signout`)
+        if(res.status==200){
+            localStorage.clear()
+            dispatch({
+                type:authConstant.LOGOUT_SUCCESS,
+            })
+        }
+        else{
+            dispatch({
+                type:authConstant.LOGOUT_FAILURE,
+                payload: {error: res.data.error}
+            })
+        }
+       
     }
 }
